requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-anthropic>=0.1.0
sqlalchemy>=2.0.0
sqlglot>=18.0.0
sentence-transformers>=2.2.2
# faiss-cpu>=1.7.4
redis>=5.0.0
boto3>=1.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
google-generativeai>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
httpx>=0.25.0
pymysql>=1.0
